updated blueprint

+++ app/controllers/events_controller.rb
@@ -11,7 +11,7 @@ class EventsController < ApplicationController
          events=Event.order(created_at: :desc).page(params[:page]).per(20)
 
         render json:{
-         events: EventBlueprint.render_as_hash(events, view: :long),
+         events: EventBlueprint.render_as_hash(events, view: :short),
          total_pages: events.total_pages,
          current_page: events.current_page
         }
